段界可以超過 64 KiB：CPU 位移改 32 位元，huge 配置每格延伸到結尾
PTO2 的矩形搬移迴圈（RE:cseg11:0x512d）是這件事的證據：它用 ESI／EDI
做 32 位元定址，**只在 EDI 變成負數時修正 selector**（es -= 8;
edi += 0x10000），往前跨過 64 KiB 邊界時完全不修。一列 640 個 byte
必然有幾列會跨過邊界——界限若真的只有 64 KiB，那幾列在真機上就會當掉。
所以 Win16 huge 配置的界限是「從我這一格到整塊的尾巴」，不是各自 64 KiB。

- cpu.Bus 的位移改成 uint32，界限交給 Bus 判（它才知道那塊有多大）；
  ModRM 的 32 位元路徑不再把位址夾成 16 位元——夾了會安靜地繞回段頭。
- 字串指令的索引與計數暫存器跟著 67 前綴走：`67 AC` 用的是 ESI。
- win16 這一側加 CPUBus 轉接層，Memory 對其他呼叫點維持 16 位元位移。
- AllocHuge 每一格的 Data 都切到整塊結尾。

TestAddr32BeyondSegment64K／TestAddr32IndexKeepsHighHalf 用執行時實際
量到的 EDI（0x101B1）當正對照。

執行從第 2,094,922 步推進到 **5,325 萬步不再撞任何缺口**，跑到步數上限。
現在卡在開場影片的等待迴圈：MCI 立刻完成，遊戲就一直重播。

+++ internal/cpu/exec.go
@@ -510,7 +510,7 @@ func (c *CPU) exec(op uint8, ip uint16) error {
 		if err != nil {
 			return c.wrap(ip, err, "位移")
 		}
-		o := operand{sel: c.dataSeg(DS), off: d}
+		o := operand{sel: c.dataSeg(DS), off: uint32(d)}
 		if op < 0xA2 {
 			v, err := c.readOp(o, sz)
 			if err != nil {
@@ -619,7 +619,7 @@ func (c *CPU) exec(op uint8, ip uint16) error {
 		frame := c.R16(SP)
 		for i := uint8(1); i < lvl; i++ {
 			c.SetR16(BP, c.R16(BP)-2)
-			v, err := c.Bus.ReadU16(c.Seg[SS], c.R16(BP))
+			v, err := c.Bus.ReadU16(c.Seg[SS], uint32(c.R16(BP)))
 			if err != nil {
 				return c.wrap(ip, err, "enter 巢狀")
 			}
@@ -671,7 +671,7 @@ func (c *CPU) exec(op uint8, ip uint16) error {
 		}
 		return c.errf(ip, "未實作的軟體中斷 INT %02Xh（AX=%04X）", n, c.R16(AX))
 	case 0xD7: // XLAT
-		v, err := c.Bus.ReadU8(c.dataSeg(DS), c.R16(BX)+uint16(c.Reg8(0)))
+		v, err := c.Bus.ReadU8(c.dataSeg(DS), uint32(c.R16(BX)+uint16(c.Reg8(0))))
 		if err != nil {
 			return c.wrap(ip, err, "xlat")
 		}
